fixed delete kafka modal open issue
