+++ src/modules/Topics/pages/Dashboard/Dashboard.tsx
@@ -3,6 +3,7 @@ import { I18nextProvider } from 'react-i18next';
 import kafkai18n from '@app/i18n';
 import { MainViewHeader } from '@app/components';
 import { FederatedProps, FederatedContext } from '@app/contexts';
+import { useModal, ModalType } from '@rhoas/app-services-ui-shared';
 
 export type DashboardProps = FederatedProps;
 
@@ -14,6 +15,16 @@ const Dashboard: React.FC<DashboardProps> = ({
   redirectAfterDeleteInstance,
   kafka,
 }) => {
+  const { showModal } = useModal<ModalType.KasDeleteInstance>();
+
+  const onDeleteInstance = () => {
+    showModal &&
+      showModal(ModalType.KasDeleteInstance, {
+        kafka,
+        onDelete: redirectAfterDeleteInstance,
+      });
+  };
+
   return (
     <I18nextProvider i18n={kafkai18n}>
       <FederatedContext.Provider
@@ -23,6 +34,7 @@ const Dashboard: React.FC<DashboardProps> = ({
           handleInstanceDrawer,
           kafka,
           redirectAfterDeleteInstance,
+          onDeleteInstance,
         }}
       >
         <MainViewHeader activeTabKey={1} />
